ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "irs-rag"
version = "0.1.0"
description = "Production-ready RAG chat service for IRS.gov content"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "IRS RAG Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
    "tldextract>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "pymupdf>=1.23.0",
    "pdfminer.six>=20221105",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "orjson>=3.9.0",
    "typer>=0.9.0",
    "slowapi>=0.1.9",
    "openai>=1.3.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.11.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
]

[project.scripts]
ingest = "app.scripts.ingest:main"
rebuild-index = "app.scripts.rebuild_index:main"
eval-suite = "app.scripts.eval_suite:main"
export-snapshot = "app.scripts.export_snapshot:main"

[tool.setuptools]

[tool.setuptools.packages.find]
include = ["app*", "infra*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SIM", "S", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "W503"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing"

